Use constants for most literal strings in dataset literalization
Sequel already did this in a few places, but this does it across
the board.  Using constants is better for performance, as using
literals in methods generates new objects on each call, which
causes more garbage to be created and need collecting.

Future commits will make similar adjustments in adapters.

This has a couple of other optimizations.  First, Symbol is
tried first in Dataset#literal.  It's likely symbols are the
most literalized objects, so putting them first saves Module#===
calls.

This also uses a regexp when escaping identifiers by default.
Ruby's String#gsub creates a regexp if you give it a string as
the pattern, so this saves object creation.