Fix tinytds adapter if DB.identifier_output_method = nil
The tinytds adapter was broken in two ways if there was no
identifier_output_method set.  First, it tried to access an
uninitialized local variable (my bad).  Second, it was broken
regardless as it yielded hashes with string keys, due to a
bug in tiny_tds.  I'm disabling the optimization used when
identifier_output_method is nil, until tiny_tds fixes their
bug.

To detect this situation, an integration test that sets
identifier_output_method to nil is added.