Support specifying nanoseconds when converting a hash to a timestamp
This also contains some minor cleanup to the last commit, since
you can not assume Rational is available (as Sequel doesn't require
it and shouldn't assume date requires it on 1.8).

This also adds some basic specs for the hash and array case. There
may be some overlap with specs defined elsewhere, but that's OK.