Allow specifying timezone offset when converting a hash or array to a timestamp
I was originally going to use this for JDBC timestamps, but it turns
out you can't use it there since there is no way to tell whether the
underlying column has an offset or not.  However, I think the ability
to specify an offset is generally useful, so I'm committing this.