When starting a new dataset graph, assume existing selection is the columns to select from the current table
Before, you could select the columns to use in the graph for the
table you are graphing into the current table using the :select
option to graph.  However, there wasn't a way to specify which
columns in the current dataset you would like to use in the graph
(when initiating a dataset graph).  This makes it use an existing
selection, assuming that it can guess the column aliases to use.

If the existing selection has unqualified columns, it
automatically qualifies them so that you don't get ambiguous
column names in the SQL query.  This reuses the Qualifier
ASTTransformer to do that.