Add :graph_alias_base association option for setting base name to use for table aliases when eager graphing
Before, the alias name was always based on the association name.
However, if you have multiple models with the same association
name that are you eager graphing simultaneously, it can be difficult
to tell which association is being referred to by which alias.
This allows the user to set the base of the alias name, which should
make that easier.