Make :server option to Model#save set the shard to use
This commit stems from a bugfix in the sharding plugin.  Before,
the sharding plugin's specced behavior was that the :server option
to Model#save affected the shard the transaction was executed on,
but not the shard the insert/update query was executed on.  This
seems obviously wrong, so I decided to fix it by having the :server
option set the shard to use for the model.

While here, I decided that basic sharding functionality for model
objects should probably be built-in, so I moved Model#set_server
and some related code from the sharding plugin to the base plugin.
This means that without loading any plugin, you can now set the
shard to use when saving/deleting models.  The sharding plugin
still offers more advanced behavior, such as automatically setting
the shard on retrieval and handling associations.

I also fixed a bug where Model#set_server did not modify @this
if @this was already present, which potentially could have caused
queries to execute on the wrong shard.