Minor spec modifications for the Oracle adapter
ruby-oci8 is broken when it comes to timezone offset handling. It
returns column values with the current timezone offset that were
input with a separate timezone offset.  Work around this in the
oracle spec by being less specific.  Add a guard in one of the
prepared statements spec to work around this failure.  This
problem was only noticed now because this is the first time
I've tested the oracle adapter in standard time, and the time
value it has to load is in daylight time.

While here, use before/after :all for a few prepared statement
specs to speed up the integration tests.