Handle usage of on_duplicate_key_update in MySQL prepared statements (Fixes #404)
This fixes a general class of bug where columns was called by
the dataset literalization code when using prepared statements.
The columns call would call the prepared statement literalization
to recurse, usually leading to a SystemStackError or a
NoMemoryError.

The fix is fairly simple.  Prepared statements now have a link to
the dataset that created them, and calling columns on a prepared
statement is now delegated to the dataset that prepared it.