Make Dataset#exists use a PlaceholderLiteralString so it works with prepared statements
Before, exists just used a LiteralString, which doesn't work with
prepared statements as it does the literalization too early. By
using a PlaceholderLiteralString, the literalization is delayed
until the query is needed.

This is a minor breakage of backwards compatibility.  A
LiteralString is a String subclass, so you could use it anywhere
you could use a string.  A PlaceholderLiteralString is an
SQL::Expression subclass, so you can only really use it inside
Sequel.  I don't expect many people are using exists and not
passing the result to another Sequel method, though.