Add database type support to the mock adapter, e.g. mock://postgres
This allows you to use the mock adapter while having database-
specific behavior.  It can be used to see what SQL will be
used on a particular database, without needing to install a
database driver for it.

This implementation is fairly simple, and there could well be
some things that don't work this way, but it seems to work
fine in simple testing.