Add Model.def_column_alias for defining alias methods for columns
This is useful if you are dealing with a legacy database and want
to create friendly method names for the underlying columns.

Note that this only adds a newer setter/getter method.  This has
no effect at the dataset level, so you can't use these aliases
in filters.  If you want to do something like that, you have
to modify the identifier input/output methods.