Dataset#from no longer handles :a__b__c___d as a.b.c AS d
Everywhere else in Sequel, this is transformed to a.b__c AS d, so
Dataset#from should do the same.

This breaks backwards compatibility for people relying on the
old behavior, but I think consistency is more important here.
Three level qualified tables can be used in Microsoft SQL
Server for cross-database access, that's the known use case
affected by this change.