Fix Dataset#naked! and remove Dataset#paginate!
The default mutation method handling only copies the opts, it does
not copy other aspects.  Add a naked! method manually that unsets
the row_proc for the current dataset.

Dataset#paginate extends the cloned dataset with a module, which
wouldn't get copied to the current dataset.  As paginate is in an
extension, it's questionable whether there should be a mutation
method created for it.  I'm removing the broken method now, if
anyone desires a mutation method for it, I can add one to the
pagination extension.