Fix alter table emulation on SQLite when foreign keys reference the table being altered
When foreign keys reference a table on SQLite, the alter table
emulation doesn't work as it must delete the table being referenced
to replace it with the altered table.  Work around the issue by
turning off foreign key support in SQLite when altering tables.
This needs to be done outside of the transaction normally used,
so even alter table commands that don't require emulation will
have it turned off and back on, but that shouldn't have a negative
effect.

Change the handling of foreign keys inside the table being altered
by always checking for them.  This is necessary as foreign key
support is now turned off by default during the altering.

Thanks to Yorick Peterse for bringing this issue to my attention.