Make table_exists? better handle tables where you don't have permissions for all columns (Fixes #422)
This should reduce the potential for error if you don't have
permissions for all columns.  It should also be faster for
tables where each row is quite large.

If you don't have permissions for the table itself, you may
still get a false negative, so note that in the RDoc.

While here, only rescue DatabaseErrors, as StandardError is a
bit too broad.  It is possible this will break adapters that
don't wrap their DatabaseErrors, but at this point I'm going
to consider those as adapter bugs that should be fixed in the
adapter.