Make set_column_type not modify defaults and NULL/NOT NULL setting on MSSQL, H2, and SQLite
This adds some integration tests for this behavior, and fixes
the cases where it wasn't behaving properly.  This is tricky on
MSSQL, because defaults are stored as constraints, and you need
to drop the constraint when altering the column type.

I wanted to get this working with derby, but limitations in derby
make testing the feature difficult (can't have a NOT NULL constraint
on a column that has a NULL default), so I have the tests marked
pending there.