A couple column default related fixes
On SQLite, wrap column defaults in parens, as the SQLite parser
complains without parens in certain cases such as function calls.

On MySQL, if Sequel::CURRENT_TIMESTAMP is used as the default
value of a DateTime/Time column, use timestamp instead of datetime
as the underlying type.  This shouldn't cause any regressions,
since previously MySQL would raise an error in this case.