Fix virtual row block handling in select_map/select_order_map if block returns an array (Fixes #428)
Before, virtual row blocks were only handled correctly if they only
returned a single argument, not an array of arguments.  As arrays
of arguments are now supported, this needed to be fixed.

In addition to fixing the array handling, this allows you to provide
both regular arguments and a block argument, just like you can
for both select and order.

While here, fix a bug where an alias wasn't recognized if it was a
string, which broke the handling of SQL::Identifiers with string
values.