Refactor serialization_modification_detection to increase performance
The previous handling was correct, but not performant, as it did
a serialize/deserialize pair both for new objects and after saving.
Change the code to just clear the original values after saving
(which fixes saved objects), and just serialize early for new
objects (which fixes new objects).

To make this either, add a private serialize_deserialized_values
instance method to the serialization plugin, so that the modification
detection plugin can use it.

While here, update the RDoc for the modification detection plugin.