Make foreign key creation work on MySQL with InnoDB engine without specifying :key option
More fun with MySQL.  With the MyISAM engine, foreign key constraints
are ignored, but it accepts constraints that don't specify the columns.
The InnoDB engine doesn't accept constraints that don't specify the
columns.  So on MySQL, you previously had to specify the :key option
to reference specific columns, even if those columns were the primary
key of the referenced table (the most common case).

This commit changes the shared mysql to parse the schema to determine
the primary keys in the referenced table, and uses those if no :key
option is specified.  It also handles self referential foreign keys
that are created at the same time as the table itself, by
introspecting the generator that created the table.