Add Database#supports_savepoints_in_prepared_transactions?, false on MySQL 5.5 (Fixes #437)
MySQL 5.5 appears not to like savepoints in prepared transactions,
which MySQL 5.1 (and I think 5.0) do not have a problem with. Add
an additional method to check the support for savepoint/prepared
transactions combination, and set it to false on MySQL 5.5+.

While here, only support prepared transactions and savepoints on
MySQL 5.0+.  I never test on MySQL < 5.0, but I don't think earlier
versions support those features.