Use Mysql#info to get accurate number of rows matched in the mysql adapter
This is kind of ugly, as it requires parsing the information out
of a string using a regular expression.  However, there does not
appear to be another way to get the information.

Unfortunately, neither do_mysql nor mysql2 appear to offer a way
to get an accurate number of rows matched.