Add a server_block extension for scoping database access inside a block to specific shards
Multiple people have asked for this over the years, but I only
recently thought of a good way to do this.  This adds some
slight refactoring to the connection pools to move the code
the transforms the given server/shard name into the actual
server/shard name into it's own method (named pick_server).

The extension then overrides the method so it will use
a different default server if one has been defined and no
specific shard has been set.  To handle nested of blocks,
the different default servers are stored in an array
(thread-local in the threaded case), which is used as a
stack with the current default as the last entry.