Add arbitrary_servers extension for connecting to arbitrary servers
With this extension, you just pass a hash of connection options that
get merged into the default options.  It's designed to work well
with the server block plugin, making it possible for to use arbitrary
servers easily, such as per-request in a web framework.

As you can't be sure if the arbitrary server will be needed again
do not cache the connection, disconnect as soon as the connection
is finished being used.