Add #all_connections method to the connection pools, which yields all available connections in the pool
This can be used to run commands on all available connections
in the pools.  Note that for threaded pools, it does not yield
connections allocated to other threads.  Also, for threaded
pools, it holds the mutex while yielding all the connections,
which basically freezes the pool while the method runs. The
threaded pools do not use recursive mutexes, so any attempt
to operate on the pool by the thread calling all avaiable
connections will probably result in an error.  Users of this
method should only operate on the underlying connection objects,
not on the pool itself.