Log full SQL string by default for prepared statements created automatically by model prepared_statements* plugins
This adds the ability to turn on logging of SQL when executing
prepared statements.  Previously, just the name was logged, along
with the arguments.  However, without full logging of the SQL, it
can be difficult to determine which query was executed, especially
if the user did not manually prepare the query.

I recommend that log_sql be set to true by default by any plugins
that automatically prepare statements.

This commit also changes the logged messages to change in some
adapters for prepared statements, where the preparation is done
by method instead of by executing SQL.