Escape identifiers when quoting on MySQL and SQLite
Before, MySQL and SQLite didn't escape identifiers when quoting,
which meant they didn't correctly handle backticks in identifier
names.

While here, fix PostgreSQL's primary key and sequence parsing
SQL to that it escapes the table and schema names, which otherwise
didn't handle embedded apostrophes correctly.  While doing that,
format the SQL nicer so that it doesn't take up so much space in
the logs.

I don't consider either of these to be security issues, since if
you allow your identifiers to be controlled by the user, you've
already lost.

Microsoft doesn't appear to offer a way to escape the ] character
in identifier.  It's possible to work around that on a case by
case basis by quoting the identifier with "" instead of [], but
that requires QUOTED_IDENTIFIER be set to ON, and Sequel can't
ensure that.

Escaping " in identifiers doesn't appear to work on Oracle either.