Allow the PG_NAMED_TYPES and PG_TYPES constants to be pre-defined
This is for future compatibility with extensions that want to
provide handlers for types, before the postgres adapter is loaded.
Because the Postgres::Database object gets the conversion procs on
first connection, if you load extensions that modify PG_TYPES
after the Database has made it's connection, it won't get those
types.  So allows extensions to be loaded before fixes the
situation.

Extensions should use the same code as here (with defined?) so
that they can work correctly if loaded both before or after
the postgres adapter is loaded.