Add pg_array extension, for dealing with string and numeric PostgreSQL arrays
This extension currently only handles the parsing and literalizing
of PostgreSQL arrays.  Unlike some other parsers, this one can
handle both single and multi-dimensional numeric and string arrays.
The previously recommended Sequel extension, sequel_column_type_array,
only handled single dimensional integer and varchar arrays. The
best parser I saw for ActiveRecord, activerecord-postgres-array,
doesn't handle multi-dimensional string arrays at all, and was
regexp-based with obvious corner cases.

There's actually two parsers in this extension.  For integer/float
arrays, we just preprocess and pass to a JSON parser, which is
generally written in C and quite fast.  For string/decimal arrays,
a hand-coded parser that was translated from node-postgres javascript
code is used.  It's slow, but should be accurate.  The slow parser
is used for decimal code as I don't believe JSON supports arbitrary
precision decimals.

The literalization part is much easier.  Instead of using the {}
format, the ARRAY[] format is used, since that allows Sequel to
reuse the existing literalization code.

This extension hooks into the native postgres adapter, registering
conversion procs so that PostgreSQL arrays are returned as PGArray
subclasses.  It can still be used with other adapters, but you'll
need to do the string->PGArray parsing manually.

PGArray is a DelegateClass of Array instead of a subclass, as there
are too many parts of Sequel that treat Array instances specially.
You can call to_a on the PGArray to get the actual array instance.

This extension will be expanded in the future. Currently, it
doesn't handle schema parsing or type casting, so using PostgreSQL
arrays at the model level is a little problematic.  That will
probably come soon.

There is also no current support for easily calling array operators.
That will probably be added in a separate extension later.

Currently, the adapter specs that test this code are only run on
the native postgres adapter.  I'll probably change that in the
future so that only the parts that test the conversion procs are
restricted to the native postgres adapter.

While here, do some minor spec cleanup.