Add schema parsing/typecast support to pg_array extension to make it easy to use with models
Without this, the schema parser assumes integer[] columns are
integer, so assigning an array of integers will raise an error
(since you can't convert an array to integers).

While here, fix the postgres specs for arrays so they run on
the non-native adapters, limiting the automatic parsing tests
to just the native adapter.