Add Model.cache_anonymous_models accessor so you can disable the caching of classes created by Sequel::Model()
This hidden caching causes weird spec glitches, since there's no
easy way to clear the cache.  It's better to just disable the
caching completely during the test suite, since you want models
to be independent.

The only time this is needed is when you are doing code reloading,
as it avoids the superclass mismatch error.  For backwards
compatibility, it will remain on by default.

There's some minor spec fallout for poorly coded specs, so clean
that up.