Convert Java::OrgPostgresqlUtil::PGobject instances to ruby strings in jdbc/postgres type conversion
This is going to be used shortly for the hstore support, but it's
possible it will have unforseen effects.  For example, if other
PostgreSQL types are returned as PGobject and the user is expecting
the Java objects instead of a ruby string.  This can be considered
as breaking backwards compatibility.

If you want to ignore this type conversion, you can just turn off
type conversion entirely for the dataset using the convert_types=
method.