Add pg_hstore extension, for dealing with the PostgreSQL hstore (key/value table) type
hstore is an extension that ships with PostgreSQL that allows you
to store a key/value table inside a single column, where keys and
values are both strings (well, values can be NULL).

The implementation is similar to the pg_array extension, using a
DelegateClass approach, with a hash used for backing.  Since
hstore only uses string keys and values, we need to override a lot
of methods to force input keys/values to strings appropriately.

Just like pg_array, add a DatasetMethods module for handling
schema parsing/typecasting, and bound argument support in the
native adapter.

Thanks to pvh's sequel-hstore for the idea behind the parser,
though I significantly changed the implementation.