Add Sequel::SQL::Wrapper class for easier dealing with wrapper objects
This is mostly to make writing extensions easier.  The sql_expr
extension needed something similar, but based it in on the
ComplexExpression class.  The SQL::Wrapper implementation was
mostly extracted from the Postgres::PGArray class, and is
fairly simple, just holding a single wrapped value.

Even though this class won't be used much in the core, it's helpful
to have in the core because then we can support it in the AST
transformers, which means that it will work with Dataset#qualify
and #unbind.

One small change with this commit is that nil.sql_expr is no
longer treated as boolean.  This is slightly backwards incompatible,
though it shouldn't break anything as it now responds to a superset
of the methods it supported before.  The reason for this change
is that NULL does not necessarily represent a boolean, it can
represent any column.  For example, you should be able to do:

  nil.sql_expr + 1 # (NULL + 1)

which the previous code didn't allow.  Now that example doesn't
make much sense, but consider:

  hash[key].sql_expr + 1

This is a case where you probably do want the "NULL + 1" SQL as
opposed to Sequel raising an exception.

I completely rewrote the sql_expr specs, since they were testing
for specific objects instead of for SQL output.  I think testing
for SQL output is more appropriate given how the objects will
be used.