Double speed of Model#new with empty hash, and quadruple speed of Model#set with empty hash
This just does a fast exit in the case that the hash is empty.
Model#initialize_set is supposed to be called for all new model
objects, so that couldn't be skipped.  This does call Hash#empty?
twice in the new? case, but if the hash is empty, the earlier check
makes it about 20% faster, and if the hash is not empty, there's
the extra call does not cause a significant slowdown.