Add schema_caching extension, to speed up loading a large number of models by loading cached schema information from a file
This idea is taken from ActiveRecord, which will be adding the
feature in 4.0.  Basically, when loading a large number of
models, there may be significant time spent loading the table
schema for each table from the database, especially in high
latency situations.

To work around this, you can cheat by caching the information to
a file, and then loading the cached file instead of querying the
database for each table/model, which is what this extension does.

To make it easy to generate the cached schema files, an -S option
is being added to bin/sequel that creates the cached schema dump
file.  It's up to the application to use load_schema_cache to
preload the schema cache.