Database#uri and #url now return nil if a connection string was not used when connecting (Fixes #453)
Before, Sequel attempted to create a URI using the the Database's
options, which only worked in simple cases.  Now, the URI given
when connecting is used verbatim, or nil is returned.

This also modifies inspect to include the URI and/or connection
options given when connecting, so it's easy to see exactly which
options were used.

This breaks backwards compatibility for Database#uri and #url.
Before, they would either return a string or raise an exception,
now they can return nil or a string, but never raise an exception.