Add null_dataset extension, for creating a dataset that never issues a database query
This is another feature that ActiveRecord will be adding in 4.0.
Basically, it's the null object pattern for datasets, so you can
create a dataset that never returns any rows (or issues other
queries).

This is easily implemented as an extension, where Dataset#nullify
just extends the cloned dataset with a module.

Note that there may be corner cases on some adapters when using
non-standard dataset methods.  Those can be dealt with on a
case by case basis.