Split pretty_table extension into two extensions
This extension is being split in two so that it can be used by
core Sequel code without adding Dataset#print.  Also,
PrettyTable.string has been added to get the string to print,
which there previously was not a way to do.

As the split extension (_pretty_table) is for internal use,
this won't be documented on the plugins page).