Make SQLite's explain use PrettyTable and return a string
Previously, it returned an array, but for compatibility with the
postgres adapter, it now returns a string.  This breaks backwards
compatibility, but I think it's a worthwhile change.

Also for compatibility, explain now accepts an argument, which is
currently ignored.  Reading SQLite explain output isn't easy, but
at least it is possible to pick out what queries are using an index
and what queries are not.