Fix Dataset#group_rollup/#group_cube on Microsoft SQL Server 2005
This adds some branching to select_group_sql to deal with the three
cases:

1) no ROLLUP/CUBE
2) standard ROLLUP/CUBE support using ROLLUP()
3) nonstandard ROLLUP/CUBE support using WITH

Previously MySQL just overrode select_group_sql to use WITH ROLLUP,
so just have it use the standard code, setting the flag to turn
on the correct syntax.

As SQL Server will be removing WITH ROLLUP/CUBE in a later version,
only use the WITH ROLLUP/CUBE code on SQL Server 2005, where
ROLLUP()/CUBE() was not supported.