Add Database#create_join_table and #drop_join_table for easily creating simple join tables for models
This is another feature that will be in ActiveRecord 4.
Unfortunately, the Sequel API is not as nice, as Sequel core does not
have any inflections support.  So there is no way to infer the key
names from the table, since tables are generally plural and keys are
generally singular (or at least that's the default for Sequel::Model
join tables).

This is the simplest API that I can think of that allows it, taking
a hash with foreign key column symbols pointing to table name
symbols:

  create_join_table(:cat_id=>:cats, :dog_id=>:dogs)

While the API is not as nice as just listing the table names,
Sequel's implementation has the following advantages:

1) Real foreign keys (via REFERENCES)
2) A composite primary key of the two foreign keys
3) An additional composite index reversing the order of the keys.

The composite primary key and additional index makes it so that most
common join table queries can benefit from one of the two indexes.

While here, add support to the migration reverser to use
drop_join_table in the down block if create_join_table is used in the
change block.