The :on_update and :on_delete entries for foreign_key can now take strings, which are used literally
This simplifies the logic so that instead of using a case
case statement, the argument is always converted to a string and
processed.  The only behavior changes are for callers that didn't
use one of the symbols previously supported, which resulted in
NO ACTION being used.  The previous behavior is likely to swallow
errors that should be caught, so it should be changed.

Since they are no longer needed, this removes the Database constants
CASCADE, NO_ACTION, SET_DEFAULTS, SET_NULL, and RESTRICT.  That
breaks backwards compatibility too, but nobody should be relying
on those constants.

While here, add a separate method for on_update_clause, simply
because having both :on_delete and :on_update use on_delete_clause
looks wrong.