Model#set_fields and #update_fields now respect :missing=>:skip and :missing=>:raise options for handling missing values
These new options allow set_fields and update_fields to handle more
cases.  :skip is probably more useful of the two options, since it
allows the methods to handle situations previously handled by
set_only and update_only, with a much simpler implementation.

To ease use with web frameworks that use hashes with strings, if
a symbol is given and doesn't exist in the hash, check for the
string version and use that if present.