Add Dataset#to_hash_groups and #select_hash_groups for getting a hash with arrays of matching values
I've needed something like this in multiple projects, and I suspect
it's a reasonably common need that people just use a manual loop
for currently.  This allows you to easily categorize a dataset by a
given column(s), with the value of that column(s) as the hash keys,
and an array of all matching values as the hash values.