Add Database#foreign_key_list for getting foreign key metadata for a given table on SQLite, MySQL, and PostgreSQL
This is the first step in allowing the schema_dumper extension
to dump foreign key constraints when dumping tables.  By default,
only require that the absolute basics, which are the referencing
columns, the referenced table, and the referenced columns.  On
adapters that make the information available, try to get the
constraint name, on delete/update actions, and deferrable
setting.  On SQLite, the referenced columns are not given if
they are the same as the primary key, so make that part of the
API.  This doesn't matter in real life, because when recreating
the constraint, if you provide nil for the referenced columns,
it assumes the primary key.

While adding this, I found that the previous code did not copy
composite foreign key constraints, so fix the code so it does.
Also, I discovered why jdbc/sqlite raises exceptions for
certain PRAGMA calls, it does so when no rows are returned, so
parse the error message string and handle that both for
foreign_key_list and indexes.