Use bigint type instead of integer for Bignum generic type on SQLite
I originally chose to use integer for Bignum on SQLite way back in
2009, but the reasons for doing so are no longer clear to me, and
the description of why it was done no longer makes sense.

The current adapters for sqlite handle this fine except for swift,
and that's due to an issue with dbic++ that I've already filed an
issue for.