Add Model#freeze implementation to get a working frozen model object
Before, calling #freeze on a model instance broke a whole bunch
of things, since model instances cache a bunch of things as instance
variables.  This attempts to make it so #freeze gives you a working
read-only object.

Note that this only handles the base and associations plugins.
There are probably plugins that ship with Sequel that are broken
on frozen objects.  Those will be handled on an as-needed basis.