Add select_remove extension which adds Dataset#select_remove for removing columns/expressions from a dataset selection
I don't think this will be used often enough to go into core
Sequel, and it also has some corner cases that can't be handled
correctly.  However, some people have been asking for it, and it's
not the simplest thing to get correct by hand even for the cases
it does handle correctly, so I think shipping this extension with
Sequel makes sense.