Add spec to check that core extensions aren't used inside Sequel
Since 3.0, it's been Sequel's policy to not rely on the core
extensions internally, so that Sequel can be used safely with the
SEQUEL_NO_CORE_EXTENSIONS setting.  However, occassionally
such usage has snuck in, and it's generally not discovered unless
someone hits the code path with the core extensions disabled.

This spec will scan all of the Sequel files looking for core
extension usage, and will fail if it detects any.

This spec actually caught usage of the core extensions in the
shared PostgreSQL adapter that had been there for close to three
years.