Increase speed of Model#delete and #destroy by about 75% for models with simple datasets
This makes Model#delete and #destroy use a similar optimization as
Model.primary_key_lookup, caching the entire SQL query to use except
for the primary key.  Since no equivalent for fetch_rows exists for
delete, add a Dataset#with_sql_delete method that just calls the
private execute_dui method.  This is used so there is no need to
clone the model's dataset to execute the SQL.

While here, do some minor micro-ops like caching certain calls in
local variables.

Only significant fallout from this is in the instance_filters plugin,
so modify that not to use this optimization if an instance filter is
being used.