Make sure Sequel::LiteralString is not missing methods if the core extensions are disabled
These methods/modules are only defined/included in String if the
core extensions are enabled.  If they are not enabled, these
methods/modules would be missing from LiteralString, which was
not intended.

Handle this by checking if the core extensions are disabled
and manually defining/including the methods/modules for
Sequel::LiteralString.

No automated tests for this code, as we don't have a test suite
that runs with the core extensions disabled.