All of Sequel's core extensions now have equivalent methods defined on the Sequel module
While the core extensions that Sequel does add by default are fairly
minor compared to other popular ruby database libraries, and are
specifically related to Sequel's DSL, it would be best if Sequel
were more usable with the core extensions disabled.  This commit
is a step towards making Sequel easy to use without the core
extensions.

This adds equivalent methods for all of the core extensions to
the Sequel module.  For example:

  :a.identifier      -> Sequel.identifier(:a)
  :a.qualify(:b)     -> Sequel.qualify(:b, :a)
  :a.sql_function(2) -> Sequel.function(:a, 2)
  {:a=>1}.case(2)    -> Sequel.case({:a=>1}, 2)

In addition, a Sequel.expr method has been added that acts as a
replacement for the sql_expr extension:

  1.sql_expr    -> Sequel.expr(1)

While using the methods defined on the main Sequel module are is
a bit more verbose than using the core extensions, it's still
reasonably concise, and much better than the previous alternative
of trying to create the objects manually.

For easy use of these builder methods in external code, they are
contained in two modules, Sequel::SQL::Builders (most methods) and
Sequel::SQL::OperatorBuilders (operators like +, &, ~).