Make the migrators only use transactions by default if the database supports transactional DDL
Using transactions by default doesn't harm things in most cases,
but it can in some cases.  For example, on SQLite, when emulating
alter_table on a table referenced by foreign keys, you need to
disabled foreign key use outside of a transaction.  If the
migrator runs the migration inside a transaction, it messes up
the foreign key reference to point to the backup table.

Seems like PostgreSQL, MSSQL, and Derby support transactional
schema, so this doesn't change things for those databases.

This adds the transaction method to the migration DSL, so you
can force transactions on a per migration basis for databases
that don't support transactional schema.