Don't assign instance variables to Java objects, for future JRuby 2.0 support (Fixes #466)
Soon to be gone are the days where you could assume everything is an
ruby object in JRuby.  In JRuby 2.0, some objects will be more equal
than others.

Sequel used to modify the connection objects to give them a
per-connection hash of native prepared statements, because that's
how Sequel works on other adapters.  However, because JRuby 2.0
will make object proxy caching opt-in, you either have to manually
set __persistent__ = true for each connection object class, or
you have to not modify that class.  Since java object proxy caching
apparently hurts performance (which is the reason they want to
make it opt-in), switch to using a single hash in the Database object,
keyed by connection (protected by a mutex for thread-safety).