Fix possible thread safety issues in creating Database objects and anonymous model classes
Creating a Database object adds the instance to the Sequel::DATABASES
array, and creating an anonymous model class using Sequel::Model()
adds an entry to the Sequel::Model::ANONYMOUS_MODEL_CLASSES hash.
Older versions of JRuby (and all versions of MRI) don't have any
issues with the previous code.  MRI has a GVL, and older versions
of JRuby had enough locking so that I was never able to construct
a failing test, even in heavily threaded situations.

However, with JRuby 1.6.7, I was finally able to get a program that
would at least fail occassionally with a ConcurrencyError when
creating Sequel::Database objects with a block in a tight loop.

It's fairly easy to protect against this using a simple global mutex,
which is what this commit does.  Any code that accesses those two
mutable global data structures should use Sequel.synchronize to
protect access.

It is possible there are other similar thread-safety corner cases
in Sequel, in very difficult to hit situations.  I'll try to be
proactive in dealing with these as I discover them.